agrega mas datos a los fixtures

+++ api/migrations/sqls/20150302145350-add-proc-benchmark-table-graphic-up.sql
@@ -9,14 +9,14 @@ $BODY$
       SUM(CASE WHEN item=1 THEN val ELSE 0 END) AS val1,
       SUM(CASE WHEN item=2 THEN val ELSE 0 END) AS val2
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (1,2)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -30,14 +30,14 @@ $BODY$
       SUM(CASE WHEN item=3 THEN val ELSE 0 END) AS val3,
       SUM(CASE WHEN item=4 THEN val ELSE 0 END) AS val4
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (3, 4)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -51,14 +51,14 @@ $BODY$
       SUM(CASE WHEN item=5 THEN val ELSE 0 END) AS val5,
       sum(1) AS total_count
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (5)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -72,14 +72,14 @@ $BODY$
       SUM(CASE WHEN item=6 THEN val ELSE 0 END) AS val6,
       sum(1) AS total_count
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (6)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -93,14 +93,14 @@ $BODY$
       SUM(CASE WHEN item=7 THEN val ELSE 0 END) AS val7,
       SUM(CASE WHEN item=8 THEN val ELSE 0 END) AS val8
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (7, 8)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -114,14 +114,14 @@ $BODY$
       SUM(CASE WHEN item=9 THEN val ELSE 0 END) AS val9,
       SUM(CASE WHEN item=10 THEN val ELSE 0 END) AS val10
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (9, 10)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -135,14 +135,14 @@ $BODY$
       SUM(CASE WHEN item=11 THEN val ELSE 0 END) AS val11,
       SUM(CASE WHEN item=12 THEN val ELSE 0 END) AS val12
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (11, 12)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -156,14 +156,14 @@ $BODY$
       SUM(CASE WHEN item=12 THEN val ELSE 0 END) AS val12,
       SUM(CASE WHEN item=13 THEN val ELSE 0 END) AS val13
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (12, 13)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -177,14 +177,14 @@ $BODY$
       SUM(CASE WHEN item=14 THEN val ELSE 0 END) AS val14,
       SUM(CASE WHEN item=15 THEN val ELSE 0 END) AS val15
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (14, 15)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -201,14 +201,14 @@ $BODY$
       SUM(CASE WHEN item=14 THEN val ELSE 0 END) AS val14,
       SUM(CASE WHEN item=15 THEN val ELSE 0 END) AS val15
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (12, 13, 14, 15)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -222,14 +222,14 @@ $BODY$
       SUM(CASE WHEN item=18 THEN val ELSE 0 END) AS val18,
       SUM(CASE WHEN item=16 THEN val ELSE 0 END) AS val16
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (18, 16)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
@@ -243,14 +243,14 @@ $BODY$
       SUM(CASE WHEN item=21 THEN val ELSE 0 END) AS val21,
       SUM(CASE WHEN item=22 THEN val ELSE 0 END) AS val22
       FROM(
-        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS INTEGER) AS val, r.updated_at AS updated_at
+        SELECT v.item AS item, CAST(coalesce(v.value, '0') AS NUMERIC) AS val, r.updated_at AS updated_at
         FROM values v
         INNER JOIN fields f ON f.id = v.id_field
         INNER JOIN reports r ON r.id = f.id_report
         INNER JOIN projects p ON p.id = r.id_project
         WHERE item IN (21, 22)
           AND r.id_template = 1 AND (r.updated_at::date BETWEEN startDate AND endDate)
-          AND CAST(coalesce(v.value, '0') AS INTEGER) > 0 AND p.name = projectName
+          AND CAST(coalesce(v.value, '0') AS NUMERIC) > 0 AND p.name = projectName
       ) AS QUERY1
       GROUP BY theDate
     ) AS QUERY2
